Put encapsulation helpers in extra files
We would like to use them elsewhere...
